# docs/swagger.yaml
basePath: /api
definitions:
  requests.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  requests.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  requests.CreateSuggestionRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
...
  requests.UpdateCartItemRequest:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
  requests.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A RESTful API for a food delivery application
  title: Food Delivery API
  version: "1.0"
paths:
  /admin/categories:
    post:
      consumes:
      - application/json
      description: Create a new category (Admin only)
      parameters:
      - description: Category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Create category (Admin only)
      tags:
      - Admin - Categories
  /admin/categories/{id}:
    delete:
      description: Delete a category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Delete category (Admin only)
      tags:
      - Admin - Categories
    put:
      consumes:
      - application/json
      description: Update a category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Update category (Admin only)
      tags:
      - Admin - Categories
...
  /admin/suggestions:
    get:
      description: Get all submitted suggestions (Admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Get all suggestions (Admin only)
      tags:
      - Admin - Suggestions
...
  /cart:
    get:
      description: Get the current user's shopping cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Get user cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Add a product to user's cart
      parameters:
      - description: Cart item data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/{id}:
    delete:
      description: Remove a product from user's cart
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Item removed from cart
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update the quantity of an item in user's cart
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated cart item data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
  /categories:
    get:
      description: Get all categories with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: List all categories
      tags:
      - Categories
...
  /suggestions:
    post:
      consumes:
      - application/json
      description: Submit a suggestion or feedback
      parameters:
      - description: Suggestion data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.CreateSuggestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Create a suggestion
      tags:
      - Suggestions
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"